feat: homeWork 관련 클래스 구조 개선 및 모달 기능 통합
